Move current contractsAPI to deprecated namespace
