+++ packages/runtime/lifecycle/src/generic/runtime.ts
@@ -4,7 +4,7 @@ import { WalletAPI } from "@marlowe.io/wallet/api";
 import { FPTSRestAPI, RestClient } from "@marlowe.io/runtime-rest-client";
 
 import { mkPayoutLifecycle } from "./payouts.js";
-import { mkContractLifecycle } from "./contracts.js";
+import { mkContractLifecycle } from "./deprecated-contracts.js";
 import { mkApplicableActionsAPI } from "./applicable-actions.js";
 
 export function mkRuntimeLifecycle(
@@ -12,12 +12,20 @@ export function mkRuntimeLifecycle(
   restClient: RestClient,
   wallet: WalletAPI
 ): RuntimeLifecycle {
-  const contracts = mkContractLifecycle(wallet, deprecatedRestAPI, restClient);
+  const deprecatedContractAPI = mkContractLifecycle(
+    wallet,
+    deprecatedRestAPI,
+    restClient
+  );
   return {
     wallet: wallet,
     restClient,
-    contracts,
+    deprecatedContractAPI,
     payouts: mkPayoutLifecycle(wallet, deprecatedRestAPI, restClient),
-    applicableActions: mkApplicableActionsAPI(restClient, wallet, contracts),
+    applicableActions: mkApplicableActionsAPI(
+      restClient,
+      wallet,
+      deprecatedContractAPI
+    ),
   };
 }
